+++ src/main/java/com/greedy/mokkoji/api/external/kakao/KakaoAccessTokenClient.java
@@ -1,4 +1,4 @@
-package com.greedy.mokkoji.api.external;
+package com.greedy.mokkoji.api.external.kakao;
 
 import com.greedy.mokkoji.api.user.dto.resopnse.kakao.KakaoAccessTokenResponse;
 import org.springframework.cloud.openfeign.FeignClient;
