+++ src/main/java/com/greedy/mokkoji/api/user/service/kakao/KakaoSocialLoginService.java
@@ -1,7 +1,7 @@
-package com.greedy.mokkoji.api.user.service;
+package com.greedy.mokkoji.api.user.service.kakao;
 
-import com.greedy.mokkoji.api.external.KakaoAccessTokenClient;
-import com.greedy.mokkoji.api.external.KakaoUserInfoClient;
+import com.greedy.mokkoji.api.external.kakao.KakaoAccessTokenClient;
+import com.greedy.mokkoji.api.external.kakao.KakaoUserInfoClient;
 import com.greedy.mokkoji.api.user.dto.resopnse.kakao.KakaoAccessTokenResponse;
 import com.greedy.mokkoji.api.user.dto.resopnse.kakao.KakaoUserInfoResponse;
 import lombok.RequiredArgsConstructor;
